requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-environ==0.12.0
django-phonenumber-field==8.1.0
idna==3.10
phonenumbers==9.0.8
psycopg==3.2.9
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
uuid==1.30
